Rubocop Move instance_behaviour_spec.rb specs in custom_validations_spec.rb
