+++ grape.gemspec
@@ -23,11 +23,11 @@ Gem::Specification.new do |s|
   s.add_runtime_dependency 'activesupport', '>= 5'
   s.add_runtime_dependency 'builder'
   s.add_runtime_dependency 'dry-types', '>= 1.1'
+  s.add_runtime_dependency 'i18n'
   s.add_runtime_dependency 'mustermann-grape', '~> 1.0.0'
   s.add_runtime_dependency 'rack', '>= 1.3.0'
   s.add_runtime_dependency 'rack-accept'
   s.add_runtime_dependency 'zeitwerk'
-  s.add_runtime_dependency 'i18n'
 
   s.files = Dir['lib/**/*', 'CHANGELOG.md', 'CONTRIBUTING.md', 'README.md', 'grape.png', 'UPGRADING.md', 'LICENSE', 'grape.gemspec']
   s.require_paths = ['lib']
