+++ lib/grape/validations/types/build_coercer.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-
 module Grape
   module Validations
     module Types
